+++ Client graphique/text.py
